Section8-GenAI/fashion-image-description/requirements.txt
google-cloud-aiplatform>=1.60.0
google-cloud-storage>=2.14.0
Pillow>=10.0.0
matplotlib>=3.8.0
